requirements.txt
astroid==2.1.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.4
future==0.17.1
gunicorn==19.9.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
line-bot-sdk==1.8.0
mccabe==0.6.1
psycopg2==2.7.6.1
pylint==2.2.2
pytz==2018.7
requests==2.21.0
six==1.12.0
soupsieve==1.7
static3==0.7.0
termcolor==1.1.0
typed-ast==1.1.1
urllib3==1.24.1
wrapt==1.10.11
